Make sure delta-mode works all of the time
There is some kind of race-condition with the power down sequenze
that i can't completely figure out.

However, it turns out that the power up sequenze itself is enough to
make sure that the image doesn't fade out if power is removed.

There is no relevant power saving with the power down of the
SSD1680, so we just remove the power down sequenze.